+++ todolist.todo
@@ -1,0 +1,5 @@
+Proyecto Biblioteca:
+    ☐ Comprobar lógica mensajes de completado, borrar mensajes de conexión BBDD @low 
+    ☐ Comprobar que todo el código se ha comentado @low 
+    ☐ Gestión préstamos @high
+    ☐ Consulta préstamos @high
